Added support for range syntax for parser and query-eval
Discovered that range query syntax from Solr / Lucene supported on CD3.
This includes year:[2104 TO 2015] (both inclusive and exclusive).  Added
support in the parser for this and also in the conversion to ML cts
query equivalent.

Added tests in Parser-tests and query-gen-tests to test this new code
that has been entered.

committing to Ron for him to have a look.